DataBlockHashIndex: reduce index table size by change uint32 to uint16

+++ table/block_suffix_index.h
@@ -36,18 +36,18 @@ namespace rocksdb {
 
 class BlockSuffixIndexBuilder {
  public:
-  BlockSuffixIndexBuilder(uint32_t n) :
+  BlockSuffixIndexBuilder(uint16_t n) :
       num_buckets_(n),
       buckets_(n),
-      estimate_((n + 2) * sizeof(uint32_t) /* n buckets, 2 num at the end */){}
-  void Add(const Slice &suffix, const uint32_t &pos);
+      estimate_((n + 2) * sizeof(uint16_t) /* n buckets, 2 num at the end */){}
+  void Add(const Slice &suffix, const uint16_t &pos);
   void Finish(std::string& buffer);
   void Reset();
   inline size_t EstimateSize() { return estimate_; }
 
  private:
-  uint32_t num_buckets_;
-  std::vector<std::vector<uint32_t>> buckets_;
+  uint16_t num_buckets_;
+  std::vector<std::vector<uint16_t>> buckets_;
   size_t estimate_;
 };
 
@@ -56,18 +56,18 @@ class BlockSuffixIndexIterator;
 class BlockSuffixIndex {
  public:
   BlockSuffixIndex(Slice  block_content);
-  void Seek(Slice& key, std::vector<uint32_t>& bucket) const;
+  void Seek(Slice& key, std::vector<uint16_t>& bucket) const;
 
-  inline uint32_t SuffixHashMapStart() const {
-    return static_cast<uint32_t>(map_start_ - data_);
+  inline uint16_t SuffixHashMapStart() const {
+    return static_cast<uint16_t>(map_start_ - data_);
   }
 
   BlockSuffixIndexIterator* NewIterator(const Slice& key) const;
 
  private:
   const char *data_;
-  uint32_t size_;
-  uint32_t num_buckets_;
+  uint16_t size_;
+  uint16_t num_buckets_;
   const char *map_start_;    // start of the map
   const char *bucket_table_; // start offset of the bucket index table
 };
@@ -75,18 +75,18 @@ class BlockSuffixIndex {
 class BlockSuffixIndexIterator {
  public:
   BlockSuffixIndexIterator(const char* start, const char* end,
-                           const uint32_t tag):
+                           const uint16_t tag):
       start_(start), end_(end), tag_(tag) {
-    current_ = start - 2 * sizeof(uint32_t);
+    current_ = start - 2 * sizeof(uint16_t);
     Next();
   }
   bool Valid();
   void Next();
-  uint32_t Value();
+  uint16_t Value();
  private:
   const char* start_; // [start_,  end_) defines the bucket range
   const char* end_;
-  const uint32_t tag_; // the fingerprint (2nd hash value) of the searching key
+  const uint16_t tag_; // the fingerprint (2nd hash value) of the searching key
   const char* current_;
 };
